Update of database tableName and tableColumn Logic
Before this commit, when we call ADatabase::connect() we write the table layout, that is tableNames and tableColumns into the member variables of the ADatabase class. This is normally ok since the database layout can be considered 'static' (as in does not change) during normal usage. It does provide a bit of a challenge however, for the database setup. The reason is this:

We call disconnect() before moving the backup, this is neccessary because we can't move/rename an open database. The backup gets made, and we call connect() again, to re-establish the connection. At this point, however, no database exists, so the tableNames and columnNames variables are written as empty. When then, later on verification of whether the data put into the database is sane takes place, it fails because the member variables holding the table layout are empty.

This problem is adressed here.