Ensurning compatibiity for older qt versions
Some uses of the auto keyword when creating Objects derived from QObject are not compatible with older compilers, as observed when testing with the oldest still supported LTS release for ubuntu (the oldest target for compatibility). This commit adresses this and changes the syntax used for instantiating QObject-based classes

To ensure the best possible compatibility, QObjects and their subclasses should be instantiated like this:

QMessageBox box(this);

instead of:

auto box = QMessageBox(this);

This method is more widely compatible, more in line with the Qt documentation and closer to the 'standard' of how QObject based classes are instantiated.