Consolidation of date and time usage
Consolidated interfacing with Dates, Times and Datetimes into the opl and ATime/ADateTime namespaces.

ADate namespace is missing, but at the moment, all dates are handled as Qt::ISOdate or Qt::Textdate, so no need at the moment for that. Might be implemented later.

Tweaked NewFlightDialog to use the new namespace exclusivly and marked duplicate old functions in ACalc namespace as deprecated for removal after new functionality has been tested.