Guards against unsuccessful Downloads in FirstRunDialog
This commit adresses #52.

The templates that are used to create the initial database are now included in the project itself. This enables the user to choose not to download the latest revision should there be connectivity issues.

Even if the user chooses to download the latest templates, there is now a fallback mechanism to create the database from local data, in case the download fails. This leaves the user with a working application instead of being stuck at the FirstRunDialog.

Creating the database layout has been decoupled from filling in the template data.