Implementation of Alog and Message Handling
Added a new System for handling Debugging, Logging and User Notification, using a qMessageHandler  to intercept and redirect calls to qDebug, qInfo, qWarning and qCritical

These messages are then formatted and written (depending on their message type) to console and/or logfile and/or the user in the GUI. Added a logs folder to AStandardPaths, the logfiles are rotated (last 10 logs are kept).

There are a number of convenince makros available - these are mostly unchanged. The available makros are:
- DEB - Messages go to Console out. Normally, they would not go into the log file, but for early testing, logging debug to the logfile is also enabled
- LOG - Logging of important application milestones. These are logged to console and log file, but not displayed to the user.
- INFO - Written to console, log file and displayed to the user in the GUI. Use INFO if the information is important enough that the user should be made aware of this. If this is not required, use LOG.
- WARN - Same as INFO, but displays a Warning instead of an information to the user
- CRIT - Same as WARN, but with a critical Warning message.

Code is documented but some more detailed docs are in the works