Tweaked NewFlightDialog
Reworked NewFlightDialog with focus on snappiness. Opening the NewFlightDialog in App was sluggish, due to the need to retrieve data from the DB  to create QCompleters and input mapping. This is now done outside the Dialog itself. The required data is retreived on application start, and then provided to the NF Dialog when it is constructed. This makes for a much better user experience when using the app. Increased app startup time but this is preferable over sluggish behaviour once the app is running.