Initial Rework of Database Interface
Working on implementing a new Airport Class, some limitations of the current Database Interface became apparent. In order to make the interface easier to use, more performant and easier to scale, this commit is the first step of a DB Interface overhaul.

Included new classes:
Row - base class for a row (entry) in the database
TailEntry : Row
AircraftEntry : Row