Integrated AirportWidget
Integrated the airportWidget into the UI where the backupWidget used to be (moved to settings). Preliminary icon only, actual tbd.

Implemented a mechanism to better detect changes to the database and update the UI where needed. In essence, the Database::databaseUpdated signal now also contains the table(s) that were edited, allowing for a more fine grained control of what to update.

MainWindow takes advantage of this to lazily update DbCompletionData on the fly.