Updates to airport database and template data
Removed airports that don't have an IATA code from the database to keep the dataset more controlled for now. With the ability to add new airports, in the rare instances where this is needed, this can be done manually.

Added timezone data to the airportWidget and dialogs, to enable local time conversion in the future

Simplified airport database layout, edited templates to better handle NULL values