Rework of input verification started
Aim of this refactor is to simplify and centralise input verification. For this purpose, a new class DatabaseCache is created. This class caches relevant data from the database which is needed for input completion and verification. It automatically updates on database changes. 

The abstract UserInput class has the methods verify and fixup, which are implemented in subclasses for the specific type of user input.

The ValidatorProvider class provides QValidator instances for use of any line edits that may need it.