New Time Class
Added a new time class to allow values greater than 24h. The input verification of a time input assumes that an input is of the form hh:mm and represents a time of day (i.e. within 00:00 to 23:59) but the total times can exceed this value. The new Time class caters for this.