small rework of NewFlightDialog
The New Flight Dialog had a bug which prevented it from correctly storing take-off and landing count when the user input deviated from the automatically calculated time of day/night.

This rework removes the option to manually adjust wether a take-off or landing is at day or night and instead uses the automatically calculated value in conjunction with the user's total input.

This should be the standard use case for the vast majority of users and streamlines the UX.

The option to manually edit the entry will be implemented in a separate dialog which allows custom entries, but warns the user that sensibility checks are disabled when doing so. This will allow for adding edge cases (e.g. day to night transition during touch and go's) manually, while keeping the 'standard' entry simpler.