Preliminary Settings rework
Adding new Airports is now supported in the NewFlightDialog.

Overhaul of the settings class. Instead of using enums and type conversions wherever settings are read or written, there are now typed functions available to set and get setting values.

Currencies are handled exclusively in the database now instead of in the settings. Some adjusments to HomeWidget and SettingsWidget pending.