Adding an interface to query the FlightAware API
The FlightAware API can be queried to retrieve basic flight information. Given a date and a flight identifier (IATA/ICAO flight ID), this data can then be used to prepopulate a new flight entry with only minimal user input required to finish logging a flight. Querying the API requires an API key, which is stored locally.

- The FlightAwareQuery class encapsulates a network manager and handles the request to the API
- The FlightAwareFlightData class encapsulates the data received from the API
- The FlightAwareJsonParser class parses the data received from the API